Prototype C++ source documentation extractor
This is a very hacky first prototype of extracting documentation from
C++ source code comments. Clang does the parsing and even the comment-
to-documentable assignment. Very nice and helpful.

Thanks also to FFI::Platypus, Inline::C, `perldoc perlguts`, the author
of the Clang package on CPAN and https://stackoverflow.com/a/46719397.
It was a fun afternoon / evening. Thanks also to the author of cldoc [1]
from whom I got the idea of exploiting Clang and the two test files.

The goal here is not to render the documentation as static HTML pages
for once, but to build a sort of perldoc for C++ projects, displaying
the documentation of a single given source file immediately in a pager,
or exporting it as Markdown (which serves as a universal intermediate
format). The Clang Comment interface seems to be well-prepared for such
a conversion, giving rendering hints for certain commands and stuff.

[1] https://github.com/jessevdk/cldoc